% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{demonstrateTests}
\alias{demonstrateTests}
\title{Show the sequence of demonstration tests from the article}
\usage{
demonstrateTests(sigDF)
}
\arguments{
\item{sigDF}{The data frame included in the package.}
}
\value{
NULL.
}
\description{
Show the sequence of demonstration tests from the article
}
\examples{
data(sigDF)
demonstrateTests(sigDF)
}
