% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{equivalentFactor}
\alias{equivalentFactor}
\title{Run equivalence tests on an ANOVA weight for a two-level factor.
The two-level factor you are testing should be effect coded so that it has a mean of 0 and a standard deviation of 1.}
\usage{
equivalentFactor(aovOut, coefname, lleq, uleq)
}
\arguments{
\item{aovOut}{Output object from a run of lm().}

\item{coefname}{Name of coefficient to examine.}

\item{lleq}{A value between -1 and 1 indicating the lower limit of the equivalence interval.}

\item{uleq}{A value between -1 and 1 indicating the upper limit of the equivalence interval.}
}
\value{
NULL.
}
\description{
Run equivalence tests on an ANOVA weight for a two-level factor.
The two-level factor you are testing should be effect coded so that it has a mean of 0 and a standard deviation of 1.
}
\examples{
# Use effect coding on the two predictors to produce standardized
# coefficients from the aov/lm model.
sampSize = 200
testDF <- data.frame(y=rnorm(sampSize),
                     x=scale(sample.int(n=2,size=sampSize,replace=TRUE)),
                     z=scale(sample.int(n=2,size=sampSize,replace=TRUE)))

out <- aov(y ~ z * x, data=testDF) # Run a conventional ANOVA
equivalentFactor(out, "z", lleq=-0.20, uleq=0.20)
}
