% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{equivalentWeight}
\alias{equivalentWeight}
\title{Run equivalence tests on a regression weight.}
\usage{
equivalentWeight(lmOut, coefname, lleq, uleq, beta = TRUE)
}
\arguments{
\item{lmOut}{Output object from a run of lm().}

\item{coefname}{Name of coefficient to examine.}

\item{lleq}{A value between -1 and 1 indicating the lower limit of the equivalence interval.}

\item{uleq}{A value between -1 and 1 indicating the upper limit of the equivalence interval.}

\item{beta}{Use TRUE if the regression is standardized and supply values for lleq and uleq between -1 and 1.}
}
\value{
NULL.
}
\description{
The selection of upper and lower limits of the equivalence band must
accord with whether or not the regression was standardized. An unstandardized
regression (with B weights) must be accompanied by upper and lower limits that
are calibrated to fit the B weight being tested. Choose a standardized regression
(e.g., with lm.beta() from the lm.beta package) to produce beta weights so that
the upper and lower limits can be calibrated on a scale of -1 to 1.
}
\examples{
myDF <- data.frame(y=rnorm(250), x=rnorm(250), z=rnorm(250))
out <- lm(y ~ x + z, data=myDF)
equivalentWeight(lmOut=out, coefname="x", lleq=-0.20, uleq=0.20)
}
