% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{fourSimulations}
\alias{fourSimulations}
\title{Run a simulation for a specified group size using four null/equivalence techniques.}
\usage{
fourSimulations(pop1, pop2, cohD = 0.3, groupSize = 200, numReps = 2500)
}
\arguments{
\item{pop1}{A numeric vector of values from which to sample.}

\item{pop2}{A numeric vector of values from which to sample.}

\item{cohD}{A numeric value calibrated as Cohen's d and representing the bandwidth of the null interval.}

\item{groupSize}{A numeric value indicating how large a group to sample from each population. Defaults to n=200.}

\item{numReps}{A numeric value indicating how many experiments to run. Defaults to n=2500.}
}
\value{
A list with four elements: A dataframe of all results, a dataframe of TOST details, a data frame of equivalence bounds details, and a data frame of Bayesian estimation details
}
\description{
Run a simulation for a specified group size using four null/equivalence techniques.
}
\examples{
myPops <- setupTwoNormalPops(popSize=10000, effectSize=0)
outputList <- fourSimulations(myPops$pop1, myPops$pop2, cohD=0.3, groupSize=100, numReps=250)
head(outputList[[1]])
}
