% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plotSimulationResults}
\alias{plotSimulationResults}
\title{Provide a basic diagnostic plot summarizing a run of the simulation.}
\usage{
plotSimulationResults(outputList)
}
\arguments{
\item{outputList}{A four item list with a data frame for each item.}
}
\value{
NULL.
}
\description{
The plot shows the HDI from Bayesian estimation in blue and red, with black
lines showing the upper and lower extents of Cohen's d used in the simulation.
Any values outside of these lines are non-equivalence according to Bayesian
estimation. The log of the Bayes Factor odds ratio is shown on the same graph
in green with a green line plotted at log(50). Any green above the line indicates
equivalence. Black dots are plotted at the top of the plot each time the
TOST and Inference by intervals tests are passed.
}
\examples{
myPops <- setupTwoNormalPops(popSize=10000, effectSize=0)
outputList <- fourSimulations(myPops$pop1, myPops$pop2, cohD=0.3, groupSize=100, numReps=250)
plotSimulationResults(outputList)
}
