% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{setupTwoNormalPops}
\alias{setupTwoNormalPops}
\title{Create two random normal populations for use in the simulation.}
\usage{
setupTwoNormalPops(popSize = 10000, effectSize = 0)
}
\arguments{
\item{popSize}{A numeric value indicating the size of the population. Defaults to N=10000.}

\item{effectSize}{A numeric value for the mean difference between the two populations. The procedure creates standard normal populations, so this value is effectively Cohen's d.}
}
\value{
A data frame with two vectors named pop1 and pop2.
}
\description{
Create two random normal populations for use in the simulation.
}
\examples{
myPops <- setupTwoNormalPops(popSize=50000, effectSize=0.3)
}
