% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{showArticleCode}
\alias{showArticleCode}
\title{Show the code from Appendix 1 of the article.
The code resides in a Markdown file called appendix1.Rmd that is included
with the package. Click on "knit" to create
an HTML file with all of the relevant output, or copy/run
pieces of code as needed.}
\usage{
showArticleCode()
}
\value{
NULL.
}
\description{
Show the code from Appendix 1 of the article.
The code resides in a Markdown file called appendix1.Rmd that is included
with the package. Click on "knit" to create
an HTML file with all of the relevant output, or copy/run
pieces of code as needed.
}
\examples{
showArticleCode()
}
