% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sigDF.R
\docType{data}
\name{sigDF}
\alias{sigDF}
\title{Stress appraisal data for the eqvtesting package.}
\format{A data frame with 540 rows and 2 variables}
\usage{
data(sigDF)
}
\description{
A dataset containing the prices and other attributes of almost 54,000
 diamonds. The variables are as follows:
}
\details{
\itemize{
  \item SIGPRES. The pressure subscale from the Stress-in-General scale
  \item GENDER. Character vector with "Female" and "Male" values
}
}
\keyword{datasets}
