% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{tabulateSimulationResults}
\alias{tabulateSimulationResults}
\title{Provide basic summary tables summarizing a run of the simulation.}
\usage{
tabulateSimulationResults(outputList)
}
\arguments{
\item{outputList}{A four item list with a data frame for each item.}
}
\value{
NULL.
}
\description{
Provide basic summary tables summarizing a run of the simulation.
}
\examples{
myPops <- setupTwoNormalPops(popSize=10000, effectSize=0)
outputList <- fourSimulations(myPops$pop1, myPops$pop2, cohD=0.3, groupSize=100, numReps=250)
tabulateSimulationResults(outputList)
}
